+++ src/main/java/chess/controller/ChessGameController.java
@@ -6,7 +6,7 @@ import chess.domain.generator.BlackGenerator;
 import chess.domain.generator.WhiteGenerator;
 import chess.domain.player.Player;
 import chess.domain.player.Team;
-import chess.utils.ChessMap;
+import chess.view.ChessMap;
 import chess.view.InputView;
 import chess.view.OutputView;
